[skip ci] done

+++ ENIGMAsystem/SHELL/Universal_System/Extensions/VideoPlayer/videoplayer.mm
@@ -48,9 +48,9 @@ void cocoa_show_cursor() {
 
 const char *cocoa_prefer_global_windowid(const char *window) {
   unsigned long long ull = strtoull(window, NULL, 10);
-  if ([NSApp windowWithWindowNumber:ull] != nil) {
+  if ([NSApp windowWithWindowNumber:(unsigned long)ull] != nil) {
     return [[NSString stringWithFormat:@"%llu",
-      [NSApp windowWithWindowNumber:ull]] UTF8String];
+      [NSApp windowWithWindowNumber:(unsigned long)ull]] UTF8String];
   }
   return window;
 }
